move `utils.py` to `utils/common.py` and update imports
