+++ scripts/dynamic_programming.py
@@ -12,7 +12,7 @@ from simple_parsing import ArgumentParser
 from gfn.casting import correct_cast
 from gfn.estimators import LogEdgeFlowEstimator
 from gfn.losses import FMParametrization
-from gfn.utils import validate
+from gfn.utils.common import validate
 from gfn.utils.modules import Tabular, Uniform
 
 parser = ArgumentParser()
